feat: adaptive v2
