+++ src/components/Authorization/Authorization.css
@@ -1,4 +1,4 @@
-.registration {
+.authorization {
   max-width: 396px;
   width: 100%;
   display: flex;
@@ -8,35 +8,35 @@
   padding-top: 70px;
 }
 
-.registration__heading {
+.authorization__heading {
   font-weight: 500;
   font-size: 24px;
   line-height: 29px;
   margin: 0 0 40px 0;
 }
 
-.registration__logo {
+.authorization__logo {
   margin-bottom: 40px;
   width: 40px;
   height: 40px;
 }
 
-.registraion__form {
+.authorization__form {
   width: 100%;
 }
 
-.registration__placeholder {
+.authorization__placeholder {
   font-weight: 400;
   font-size: 10px;
   line-height: 12px;
   color: #a0a0a0;
 }
 
-.registration__field {
+.authorization__field {
   margin: 0 0 24px 0;
 }
 
-.registration__input {
+.authorization__input {
   width: 100%;
   margin: 0;
   padding: 0 0 10px 0;
