+++ src/components/Header/Header.css
@@ -31,10 +31,13 @@
   text-decoration: none;
   color: darkgray;
   transition: 0.2s;
+
 }
 
 .header__link:hover {
   transform: scale(1.1);
+  opacity: 0.8;
+
 }
 
 .header__link_active {
