+++ src/components/NotFound/NotFound.css
@@ -34,3 +34,16 @@
 .not-found__button:hover {
     text-decoration: underline;
 }
+
+@media screen and (max-width: 400px) {
+    .not-found__heading {
+        font-weight: 400;
+        font-size: 80px;
+        line-height: 97px;
+    }
+
+    .not-found__subheading {
+        font-size: 12px;
+
+    }
+}
