+++ src/components/Portfolio/Portfolio.css
@@ -36,7 +36,7 @@
 
 .portfolio__link:hover {
     background-color: #FD5602;
-    padding-left: 20px;
+    padding-left: 10px;
 }
 
 @media screen and (max-width: 768px) {
