--- src/utils.js
@@ -1,4 +1,0 @@
-export const PROFILE_MOCK_DATA = {
-    name: "Азиз",
-    email: "pierre-dunn@mail.ru"
-}
