+++ src/utils/utils.js
@@ -1,0 +1,33 @@
+import image from "../assets/images/movie_pic.jpg";
+export const PROFILE_MOCK_DATA = {
+    name: "Азиз",
+    email: "pierre-dunn@mail.ru"
+}
+
+export const INITIAL_VALUES = {
+    email: "",
+    password: ""
+};
+
+export const REGISTRATION_TEXTS = {
+    heading: "Добро пожаловать!",
+    button: "Зарегистрироваться",
+    linkText: "Уже зарегистрированы? ",
+    link: "Войти",
+    route: "/signin"
+};
+
+export const LOGIN_TEXTS = {
+    heading: "Рады видеть!",
+    button: "Войти",
+    linkText: "Ещё не зарегистрированы? ",
+    link: "Регистрация",
+    route: "/signup"
+};
+
+export const MOCK_DATA = {
+    title: "33 слова о дизайне",
+    length: "1 ч 47 м",
+    image,
+    isSaved: false
+}
